create .game-pieces on click function
When the user clicks on a game piece, log the image's source to the variable gamePiece. If any div has the class .active applied, replace that div's (child) img's source with gamePiece and set the opacity to 1. After, strip the active class, so it doesn't get applied to more than one div at a time.